# 小工具，UI基础面板，对话，Buff收工！

+++ NEGame/Assets/EditorFramework/MenuItemEx/Editor/UnusedResourcesWindow.cs
@@ -1,5 +1,7 @@
+using System;
 using System.Collections.Generic;
 using UnityEditor;
+using UnityEditor.SceneManagement;
 using UnityEngine;
 using static UnityEngine.GUILayout;
 
@@ -8,33 +10,48 @@ namespace Shawn.EditorFramework
     
     public class UnusedResourcesWindow : EditorWindow
     {
+        AssetCollector collector = new AssetCollector();
+
         Vector2 scroll = default;
 
-        [MenuItem("Assets/工具/01.查询未引用资源")]
+        public const string Check_Folder_Path = "Assets"; //待检查目录
+
+        [MenuItem("Universal Tools/资源/01.查询未引用资源 %/", priority = -1)]
         static void CheckUnusedResource()
         {
+            EditorSceneManager.SaveScene(EditorSceneManager.GetActiveScene());
             UnusedResourcesWindow window = CreateInstance<UnusedResourcesWindow>();
+            window.collector.Collection(new string[] { Check_Folder_Path });
+            window.CopyDeleteFileList(window.collector.deleteFileList);
             window.Show();
         }
 
-        List<DeleteAssetData> deleteAssetDatas = new List<DeleteAssetData>()
-        {
-            new DeleteAssetData(),
-            new DeleteAssetData()
-        };
+        
+
+        #region Test
+        /*        List<DeleteAssetData> deleteAssetDatas = new List<DeleteAssetData>()
+                {
+                    new DeleteAssetData(),
+                };*/
+        #endregion
+
+        List<DeleteAssetData> deleteAssetCollection = new List<DeleteAssetData>();
 
         void OnGUI()
         {
             using (var horizontal = new EditorGUILayout.HorizontalScope("box"))
             {
-                EditorGUILayout.LabelField("以下资源未使用 可删除");
+                EditorGUILayout.LabelField("以下资源未使用(仅静态场景中未使用,谨慎使用删除操作)");
             }
 
             using (ScrollViewScope scrollScope = new ScrollViewScope(scroll))
             {
                 scroll = scrollScope.scrollPosition;
-                foreach (DeleteAssetData data in deleteAssetDatas)
+                List<DeleteAssetData> tempList = new List<DeleteAssetData>() ;
+                
+                foreach (DeleteAssetData data in deleteAssetCollection)
                 {
+                    tempList.Add(data);
                     if (string.IsNullOrEmpty(data.filePath)) continue;
                     {
                         using (var horizontal = new EditorGUILayout.HorizontalScope())
@@ -48,10 +65,30 @@ namespace Shawn.EditorFramework
                             }
                             if (Button("移除", Width(40), Height(20)))
                             {
-                                AssetDatabase.DeleteAsset(data.filePath);
+                                if (EditorUtility.DisplayDialog("", "是否确定删除该文件", "确定", "取消"))
+                                {
+                                    tempList.Remove(data);
+                                    AssetDatabase.DeleteAsset(data.filePath);
+                                }
                             }
                         }
                     }
+                    deleteAssetCollection = tempList;
+                }
+            }
+        }
+        private void CopyDeleteFileList(IEnumerable<string> deleteFileList)
+        {
+            deleteAssetCollection.Clear();
+            foreach (string asset in deleteFileList)
+            {
+                string filePath = AssetDatabase.GUIDToAssetPath(asset);
+
+                if (!string.IsNullOrEmpty(filePath))
+                {
+                    DeleteAssetData tempDelData = new DeleteAssetData();
+                    tempDelData.filePath = filePath;
+                    deleteAssetCollection.Add(tempDelData);
                 }
             }
         }
